[Tests] Reduced code duplication in custom providers test cases
