+++ tests/lib/MVC/Symfony/Security/User/BaseProviderTestCase.php
@@ -12,14 +12,18 @@ use Ibexa\Contracts\Core\Repository\PermissionResolver;
 use Ibexa\Contracts\Core\Repository\UserService;
 use Ibexa\Contracts\Core\Repository\Values\Content\ContentInfo;
 use Ibexa\Contracts\Core\Repository\Values\User\User as APIUser;
+use Ibexa\Core\Base\Exceptions\NotFoundException;
 use Ibexa\Core\MVC\Symfony\Security\User as MVCUser;
 use Ibexa\Core\MVC\Symfony\Security\User\BaseProvider;
 use Ibexa\Core\MVC\Symfony\Security\UserInterface;
 use Ibexa\Core\Repository\Values\Content\Content;
 use Ibexa\Core\Repository\Values\Content\VersionInfo;
 use Ibexa\Core\Repository\Values\User\User;
+use Ibexa\Core\Repository\Values\User\UserReference;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface as SymfonyUserInterface;
 
 abstract class BaseProviderTestCase extends TestCase
@@ -31,6 +35,10 @@ abstract class BaseProviderTestCase extends TestCase
     protected BaseProvider $userProvider;
 
     abstract protected function buildProvider(): BaseProvider;
+
+    abstract protected function getUserIdentifier(): string;
+
+    abstract protected function getUserServiceMethod(): string;
 
     protected function setUp(): void
     {
@@ -73,6 +81,34 @@ abstract class BaseProviderTestCase extends TestCase
         self::assertSame(['ROLE_USER'], $user->getRoles());
     }
 
+    public function testRefreshUserNotFound(): void
+    {
+        $userId = 123;
+        $apiUser = $this->buildUserValueObjectStub($userId);
+        $user = $this->createMock(UserInterface::class);
+        $user
+            ->expects(self::once())
+            ->method('getAPIUser')
+            ->willReturn($apiUser);
+
+        $this->userService
+            ->expects(self::once())
+            ->method('loadUser')
+            ->with($userId)
+            ->willThrowException(new NotFoundException('user', 'foo'));
+
+        $this->expectException(UserNotFoundException::class);
+        $this->userProvider->refreshUser($user);
+    }
+
+    public function testRefreshUserNotSupported(): void
+    {
+        $user = $this->createMock(SymfonyUserInterface::class);
+
+        $this->expectException(UnsupportedUserException::class);
+        $this->userProvider->refreshUser($user);
+    }
+
     protected function createUserWrapperMockFromAPIUser(User $apiUser, int $userId): UserInterface & MockObject
     {
         $refreshedAPIUser = clone $apiUser;
@@ -96,6 +132,50 @@ abstract class BaseProviderTestCase extends TestCase
         ;
 
         return $user;
+    }
+
+    public function testRefreshUser(): void
+    {
+        $userId = 123;
+        $apiUser = $this->buildUserValueObjectStub($userId);
+        $user = $this->createUserWrapperMockFromAPIUser($apiUser, $userId);
+
+        $this->permissionResolver
+            ->expects(self::once())
+            ->method('setCurrentUserReference')
+            ->with(new UserReference($apiUser->getUserId()));
+
+        self::assertSame($user, $this->userProvider->refreshUser($user));
+    }
+
+    public function testLoadUserByUsername(): void
+    {
+        $username = $this->getUserIdentifier();
+        $apiUser = $this->createMock(APIUser::class);
+
+        $this->userService
+            ->expects(self::once())
+            ->method($this->getUserServiceMethod())
+            ->with($username)
+            ->willReturn($apiUser);
+
+        $user = $this->userProvider->loadUserByIdentifier($username);
+        self::assertInstanceOf(UserInterface::class, $user);
+        self::assertSame($apiUser, $user->getAPIUser());
+        self::assertSame(['ROLE_USER'], $user->getRoles());
+    }
+
+    public function testLoadUserByUsernameUserNotFound(): void
+    {
+        $username = $this->getUserIdentifier();
+        $this->userService
+            ->expects(self::once())
+            ->method($this->getUserServiceMethod())
+            ->with($username)
+            ->willThrowException(new NotFoundException('user', $username));
+
+        $this->expectException(UserNotFoundException::class);
+        $this->userProvider->loadUserByIdentifier($username);
     }
 
     protected function buildUserValueObjectStub(int $userId): User
